Fix compile errors
Apparently room won't compile if the room objects are private.
