+++ app/settings.gradle
@@ -1,0 +1,1 @@
+kapt.incremental.apt=true
